+++ CHANGELOG.md
@@ -15,7 +15,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Changed
 
-- `RDF.Graph.new` now takes the graph name as a `name` option
+- the constructor functions for `RDF.Graph`s and `RDF.Dataset`s now take the 
+  graph name resp. dataset name through a `name` option, instead of the first
+  argument
 
 
 [Compare v0.5.4...HEAD](https://github.com/marcelotto/rdf-ex/compare/v0.5.4...HEAD)